requirements.txt
cryptography
pycryptodomex